Updated Week 1 Exercises and Assignments

+++ CAOAL4-ACTIVITIES/Week1/Day4/excer1_1/CargoItem.java
@@ -1,12 +1,10 @@
-import java.util.List;
-
 class CargoItem {
-    long id;
+    int id;
     String name;
     double weight;
     String destination;
 
-    public CargoItem(long id, String name, double weight, String destination) {
+    public CargoItem(int id, String name, double weight, String destination) {
         this.id = id;
         this.name = name;
         this.weight = weight;
